flattend componenet directory and refactored css for tag-manage, project-manage, project-form
