+++ src/App.scss
@@ -5,5 +5,9 @@
 @import './stylesheets/forms';
 
 @import './components/vertical-nav-bar/vertical-nav-bar';
-@import './components/forms/text-field/text-field';
-@import './components/forms/form-tabs/form-tabs';
+@import './components/text-field/text-field';
+@import './components/form-tabs/form-tabs';
+@import './components/tag-manage/tag-manage';
+
+@import './components/project-form/project-form';
+@import './components/project-manage/project-manage';
