+++ src/components/artwork-form.component.js
@@ -2,12 +2,12 @@ import React from 'react';
 import PropTypes from 'prop-types';
 
 import TextField from './text-field/text-field.component';
-import TagManage from './tag-manage.component';
-import ArtworkService from '../../services/artwork.service';
+import TagManage from './tag-manage/tag-manage.component';
+import ArtworkService from '../services/artwork.service';
 import ImageForm from './image-form.component';
-import ArtworkCard from '../artwork-card.component';
+import ArtworkCard from './artwork-card.component';
 
-import {deepMerge} from '../../helpers';
+import {deepMerge} from '../helpers';
 
 const Artwork$ = new ArtworkService();
