+++ src/components/form-tabs/_form-tabs.scss
@@ -1,4 +1,4 @@
-@import '../../../stylesheets/variables';
+@import '../../stylesheets/variables';
 
 .form-tabs{
     padding:0;
