+++ src/components/gallery.component.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types'
-import ArtworkCard from '../artwork-card.component';
+import ArtworkCard from './artwork-card.component';
 
 class Gallery extends React.Component {
     constructor(){
