+++ src/components/project-manage/_project-manage.scss
@@ -1,0 +1,13 @@
+@import '../../stylesheets/variables';
+
+.project-manage-container{
+    flex-basis: 100%;
+    padding: 1rem;
+    .layout-container{
+        margin: 0 auto;
+        display: flex;
+        justify-content:space-around;
+        position:relative;
+    
+    }
+}
