+++ src/components/project-manage/project-manage.component.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import {Route, Link} from 'react-router-dom'
-import ProjectForm from '../forms/project-form.component';
+import ProjectForm from '../project-form/project-form.component';
 import SideMenu from '../side-menu.component';
 
 class ProjectManage extends React.Component {
@@ -16,7 +16,7 @@ class ProjectManage extends React.Component {
 
     render(){
         return(
-            <div className="width-12">
+            <div className="project-manage-container">
                 <div className="layout-container">
                     <ProjectForm/>
                 </div>
