+++ src/components/text-field/_text-field.scss
@@ -1,4 +1,4 @@
-@import '../../../stylesheets/variables';
+@import '../../stylesheets/variables';
 
 .text-field-group {
 	padding:.5em 0;
