+++ src/stylesheets/_layout.scss
@@ -119,8 +119,7 @@
 
 .main-header {
 	display:flex;
-	padding: 1rem;
-	justify-content:flex-end;
+	justify-content:space-between;
 	flex-flow: wrap;
 	border-bottom: 1px solid $color-primary-grey;
 	background-color:$color-background-white;
